+++ SmellIt.Application/Features/WebsiteTexts/Commands/DeleteWebsiteTextByEncodedName/DeleteWebsiteTextByEncodedNameCommandHandler.cs
@@ -5,15 +5,24 @@ namespace SmellIt.Application.Features.WebsiteTexts.Commands.DeleteWebsiteTextBy
 
 public class DeleteWebsiteTextByEncodedNameCommandHandler : IRequestHandler<DeleteWebsiteTextByEncodedNameCommand>
 {
+    private readonly IUserContext _userContext;
     private readonly IWebsiteTextRepository _websiteTextRepository;
 
-    public DeleteWebsiteTextByEncodedNameCommandHandler(IWebsiteTextRepository websiteTextRepository)
+    public DeleteWebsiteTextByEncodedNameCommandHandler(IUserContext userContext,IWebsiteTextRepository websiteTextRepository)
     {
+        _userContext = userContext;
         _websiteTextRepository = websiteTextRepository;
     }
 
     public async Task<Unit> Handle(DeleteWebsiteTextByEncodedNameCommand request, CancellationToken cancellationToken)
     {
+        var currentUser = _userContext.GetCurrentUser();
+
+        if (currentUser == null || !currentUser.IsInRole("Admin"))
+        {
+            return Unit.Value;
+        }
+
         await _websiteTextRepository.DeleteAsync(request.EncodedName);
         return Unit.Value;
     }
